Revert "x86: normalize types used with _mm_cvtsi128_si32"
This reverts commit cbbc5cc3ab19003937bd97ebcea811961eb42736.

Reason for revert: This causes some assertion / fuzzing failures.

Bug: oss-fuzz:49590
Bug: oss-fuzz:49595

+++ aom_dsp/x86/jnt_sad_ssse3.c
@@ -53,8 +53,7 @@ unsigned int aom_sad4xh_sse2(const uint8_t *a, int a_stride, const uint8_t *b,
 
   // At this point, we have two 32-bit partial SADs at bit[0:31] and [64:95].
   const unsigned int res =
-      (unsigned int)(_mm_cvtsi128_si32(sad) +
-                     _mm_cvtsi128_si32(_mm_srli_si128(sad, 8)));
+      _mm_cvtsi128_si32(sad) + _mm_cvtsi128_si32(_mm_srli_si128(sad, 8));
 
   return res;
 }
@@ -85,8 +84,7 @@ unsigned int aom_sad8xh_sse2(const uint8_t *a, int a_stride, const uint8_t *b,
   }
 
   const unsigned int res =
-      (unsigned int)(_mm_cvtsi128_si32(sad) +
-                     _mm_cvtsi128_si32(_mm_srli_si128(sad, 8)));
+      _mm_cvtsi128_si32(sad) + _mm_cvtsi128_si32(_mm_srli_si128(sad, 8));
 
   return res;
 }
@@ -110,8 +108,7 @@ unsigned int aom_sad16xh_sse2(const uint8_t *a, int a_stride, const uint8_t *b,
   }
 
   const unsigned int res =
-      (unsigned int)(_mm_cvtsi128_si32(sad) +
-                     _mm_cvtsi128_si32(_mm_srli_si128(sad, 8)));
+      _mm_cvtsi128_si32(sad) + _mm_cvtsi128_si32(_mm_srli_si128(sad, 8));
 
   return res;
 }
@@ -137,8 +134,7 @@ unsigned int aom_sad32xh_sse2(const uint8_t *a, int a_stride, const uint8_t *b,
   }
 
   const unsigned int res =
-      (unsigned int)(_mm_cvtsi128_si32(sad) +
-                     _mm_cvtsi128_si32(_mm_srli_si128(sad, 8)));
+      _mm_cvtsi128_si32(sad) + _mm_cvtsi128_si32(_mm_srli_si128(sad, 8));
 
   return res;
 }
@@ -164,8 +160,7 @@ unsigned int aom_sad64xh_sse2(const uint8_t *a, int a_stride, const uint8_t *b,
   }
 
   const unsigned int res =
-      (unsigned int)(_mm_cvtsi128_si32(sad) +
-                     _mm_cvtsi128_si32(_mm_srli_si128(sad, 8)));
+      _mm_cvtsi128_si32(sad) + _mm_cvtsi128_si32(_mm_srli_si128(sad, 8));
 
   return res;
 }
@@ -191,8 +186,7 @@ unsigned int aom_sad128xh_sse2(const uint8_t *a, int a_stride, const uint8_t *b,
   }
 
   const unsigned int res =
-      (unsigned int)(_mm_cvtsi128_si32(sad) +
-                     _mm_cvtsi128_si32(_mm_srli_si128(sad, 8)));
+      _mm_cvtsi128_si32(sad) + _mm_cvtsi128_si32(_mm_srli_si128(sad, 8));
 
   return res;
 }
